added figure 5 g
this is a figure meant to measure the activation probability of a (state, action) before and after is was found to be a shock ( negative reward ).

To do this we let the agent explore freely without reward. It then does a series of planning to measure the activation probability before shock.

We then implant a memory of a shock to the agent. Afterwards, we then measure the activation probability of taking a particular (state, action).